Install third-party assets via npm
With this change, those assets become local files from the point of view of
assetpack. This means no downloads are done on the fly but one must invoke
npm first (which could be automated).

With this we could probably use `obs-service-node_modules` like cockpit
does (if *all* dependencies could be installed that way).

The version of some dependencies has been slightly increased with this.
However, there are no major version bumps and everything seems to work
and look as expected in my manual tests.

Related ticket: https://progress.opensuse.org/issues/153427