Pass error on worker registration to avoid it from showing up as idle
So far the current error of a worker is cleared on registration and only
updated via the status message sent via web sockets. This leaves a small
window in which a worker might appear idle when it is actually broken.

Note that this should normally not lead to wrong job assignments because
job assignments are only possible as soon as the web socket connection is
established.