WIP: Fix possibly excessive memory use when computer test result overview
When accessing the test results overview page (especially without
parameters, e.g. http://localhost:9526/tests/overview) a very huge set of
jobs is considered. This leads to a very excessive memory usage, e.g. with
current OSD data the RSS rises over 7 GiB and the request takes very long.
There's already a limit of 2000 jobs which is also effective but
unfortunately not passed to `complex_query` and only used later on. I
tracked the memory usage while running the code with debug printing it it
was very clear that `complex_query` is the culprit.

This change uses the limit from the start. This means additional
filtering done in Perl code also needs to happen as part of the now
limited initial query. So this change replaces these Perl loops/grep
with passing the filter parameters directly to the database. This
slightly changes the order of things but should lead to the same end
result.

With this change unconstrained queries (e.g. just `/tests/overview` without
parameters) show more results. The code before this change still limited
results to a certain build this way. It makes more sense for unconstrained
queries to be actually unconstrained so with this change I didn't put any
effort into preserving the old behavior. This change in behavior is also
the reason why tests needed to be adjusted.

Related ticket: https://progress.opensuse.org/issues/192448